Moved specialization of tuple_size/tuple_element for AZStd::pair/AZStd::array
The AZStd::pair class needs `std::tuple_element` to be specialized
before the declaration of the `AZStd::get` overloads for the AZStd::pair
class in order to compile using clang/gcc.
