+++ Code/Framework/AzCore/AzCore/std/tuple.h
@@ -91,64 +91,3 @@ namespace AZStd
             AZStd::make_index_sequence<AZStd::tuple_size<AZStd::decay_t<Tuple>>::value>{});
     }
 } // namespace AZStd
-
-namespace AZStd
-{
-    //! AZStd::pair to std::tuple function for replicating the functionality of std::tuple assignment operator from std::pair
-    template<class T1, class T2>
-    constexpr tuple<T1, T2> tuple_assign(const AZStd::pair<T1, T2>& azPair);
-
-    //! AZStd::pair to std::tuple function for replicating the functionality of std::tuple assignment operator from std::pair
-    template<class T1, class T2>
-    constexpr tuple<T1, T2> tuple_assign(AZStd::pair<T1, T2>&& azPair);
-} // namespace AZStd
-
-// The tuple_size and tuple_element classes need to be specialized in the std:: namespace since the AZStd:: namespace alias them
-// The tuple_size and tuple_element classes is to be specialized here for the AZStd::pair class
-// The tuple_size and tuple_element classes is to be specialized here for the AZStd::array class
-
-//std::tuple_size<std::pair> as defined by C++ 11 until C++ 14
-//template< class T1, class T2 >
-//struct tuple_size<std::pair<T1, T2>>;
-//std::tuple_size<std::pair> as defined since C++ 14
-//template <class T1, class T2>
-//struct tuple_size<std::pair<T1, T2>> : std::integral_constant<std::size_t, 2> { };
-
-//std::tuple_element<std::pair> as defined since C++ 11
-//template< class T1, class T2 >
-//struct tuple_element<0, std::pair<T1,T2> >;
-//template< class T1, class T2 >
-//struct tuple_element<1, std::pair<T1,T2> >;
-
-namespace std
-{
-    // Suppressing clang warning error: 'tuple_size' defined as a class template here but previously declared as a struct template [-Werror,-Wmismatched-tags]
-    AZ_PUSH_DISABLE_WARNING(, "-Wmismatched-tags")
-    template<class T1, class T2>
-    struct tuple_size<AZStd::pair<T1, T2>> : public AZStd::integral_constant<size_t, 2> {};
-
-    template<class T1, class T2>
-    struct tuple_element<0, AZStd::pair<T1, T2>>
-    {
-    public:
-        using type = T1;
-    };
-
-    template<class T1, class T2>
-    struct tuple_element<1, AZStd::pair<T1, T2>>
-    {
-    public:
-        using type = T2;
-    };
-
-    template<class T, size_t N>
-    struct tuple_size<AZStd::array<T, N>> : public AZStd::integral_constant<size_t, N> {};
-
-    template<size_t I, class T, size_t N>
-    struct tuple_element<I, AZStd::array<T, N>>
-    {
-        static_assert(I < N, "AZStd::tuple_element has been called on array with an index that is out of bounds");
-        using type = T;
-    };
-    AZ_POP_DISABLE_WARNING
-} // namespace std
